requirements.txt
asgiref==3.11.0
Django==4.2.27
gunicorn==23.0.0
packaging==26.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
typing_extensions==4.15.0
mercadopago==2.3.0
python-dotenv==1.0.0
